requirements.txt
# API / backend
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
grpcio>=1.62.0
grpcio-tools>=1.62.0

# Data & streaming
kafka-python>=2.0.2
redis>=5.0.3
snowflake-connector-python>=3.8.0
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.13.0

# ML / models
torch>=2.2.0
torchvision>=0.17.0
opencv-python-headless>=4.9.0
stable-baselines3>=2.3.0
gymnasium>=0.29.0

# Dashboard
streamlit>=1.34.0
requests>=2.31.0

# Observability
prometheus-client>=0.20.0
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0

# Load testing (test only)
locust>=2.24.0

# Testing
pytest>=8.1.0
coverage>=7.4.0
